# src/swagger/atlantida-api.yaml
openapi: 3.0.3
info:
  title: Atlântida API
  description: API para o gerenciamento de mergulhos - Plataforma Atlântida.
  version: 1.0.0
servers:
  - url: http://localhost:3000
components:
 securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/addresses/{id}:
    get:
      summary: Obter endereço por ID
      description: Obtém os detalhes de um endereço específico.
      tags:
        - addresses
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: O ID do endereço
      responses:
        '200':
          description: Detalhes do endereço
          content:
            application/json:
              schema:
                $ref: '#/definitions/Address'
    put:
      summary: Atualizar endereço
      description: Atualiza os detalhes de um endereço.
      tags:
        - addresses
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: O ID do endereço
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/definitions/Address'
      responses:
        '200':
          description: Endereço atualizado com sucesso
    delete:
      summary: Excluir endereço
      description: Exclui um endereço.
      tags:
        - addresses
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: O ID do endereço
      responses:
        '204':
          description: Endereço excluído com sucesso
...
  /certificates/{id}:
    get:
      summary: Busca certificado por id
      tags:
        - certificates
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Certificado encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Certificate'
    put:
      summary: Atualizar certificado
      tags:
        - certificates
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Certificate'
      responses:
        '200':
          description: Certificado atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Certificate'
    delete:
      summary: Deletar certificado
      tags:
        - certificates
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Certificado deletado com sucesso
...
  /api/diveLogs/{id}:
    get:
      summary: Busca mergulho por id
      tags:
        - dive logs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mergulho encontrado
    put:
      summary: Atualizar mergulho
      tags:
        - dive logs
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/definitions/DiveLog'
      responses:
        '200':
          description: Mergulho atualizado com sucesso
    delete:
      summary: Deletar mergulho
      tags:
        - dive logs
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Mergulho deletado com sucesso
...
  /api/comments/{id}:
    get:
      summary: Filtrar por id
      tags:
        - comments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comentário encontrado
    put:
      summary: Atualizar comentário
      tags:
        - comments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/definitions/Comment'
      responses:
        '200':
          description: Comentário atualizado com sucesso
    delete:
      summary: Deletar comentário
      tags:
        - comments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Comentário deletado com sucesso
definitions:
  Address:
    type: object
    required:
      - country
      - state
      - city
      - neighborhood
      - street
      - number
      - postalCode
    properties:
      country:
        type: string
        description: O país do usuário. (e.g., Brasil)
      state:
        type: string
        description: O estado do usuário dentro do país. (e.g., SP)
      city:
        type: string
        description: A cidade do usuário.
      neighborhood:
        type: string
        description: O bairro do usuário. (opcional)
      street:
        type: string
        description: O endereço da rua do usuário.
      number:
        type: integer
        description: O número do endereço do usuário.
      complement:
        type: string
        description: Informações adicionais de endereço (e.g., número do apartamento, andar). (opcional)
      postalCode:
        type: string
        description: O código postal do usuário.
...
  DiveLog:
    type: object
    required:
    - title
    - location
    - date
    - type
    - depth
    - bottomTimeInMinutes
    - userId
    properties:
      title:
        type: string
        description: Título do mergulho
      location:
        type: string
        description: Localização do mergulho
      date:
        type: string
        format: date-time
        description: Data e hora do mergulho
      type:
        type: string
        description: Tipo de mergulho (por exemplo, Scuba Diving)
      depth:
        type: number
        format: float
        description: Profundidade máxima do mergulho em metros
      bottomTimeInMinutes:
        type: integer
        description: Tempo de permanência na água em minutos
      waterType:
        type: string
        description: Tipo de água (Salgada ou Doce)
      waterBody:
        type: string
        description: Corpo de água (por exemplo, Oceano, Lago, Pedreira, Rio ou Outro)
      weatherConditions:
        type: string
        description: Condições meteorológicas durante o mergulho
      temperature:
        type: object
        description: Temperaturas em diferentes partes do corpo e da água
        properties:
          air:
            type: number
            format: float
            description: Temperatura do ar em graus Celsius
          surface:
            type: number
            format: float
            description: Temperatura da superfície da água em graus Celsius
          bottom:
            type: number
            format: float
            description: Temperatura do fundo da água em graus Celsius
      visibility:
        type: string
        description: Visibilidade da água
      waves:
        type: string
        description: Condições das ondas
      current:
        type: string
        description: Condições da corrente
      surge:
        type: string
        description: Condições de surge
      suit:
        type: string
        description: Tipo de traje usado
      weight:
        type: string
        description: Peso do traje e equipamento
      additionalEquipment:
        type: array
        items:
          type: string
        description: Equipamentos adicionais usados
      cylinder:
        type: object
        description: Informações sobre o cilindro de ar usado
        properties:
          type:
            type: string
            description: Tipo do cilindro (por exemplo, Aluminum)
          size:
            type: number
            format: float
            description: Tamanho do cilindro em litros
          gasMixture:
            type: string
            description: Mistura de gases no cilindro (por exemplo, Air)
          initialPressure:
            type: number
            format: float
            description: Pressão inicial do cilindro em bar
          finalPressure:
            type: number
            format: float
            description: Pressão final do cilindro em bar
          usedAmount:
            type: number
            format: float
            description: Quantidade de gás usado em litros
      rating:
        type: string
        description: Avaliação do mergulho
      notes:
        type: string
        description: Observações sobre o mergulho
      buddy:
        type: string
        description: Nome do companheiro de mergulho
      photos:
        type: array
        items:
          type: object
          properties:
            data:
              type: string
              format: byte
              description: Dados da foto codificados em base64
            contentType:
              type: string
              description: Tipo de conteúdo da foto (por exemplo, image/jpeg)
      userId:
        type: string
        description: Identificador do usuário que registrou o mergulho
...
  Comment:
    type: object
    required:
    - rating
    - userId
    - divingSpotId
    properties:
      rating:
        type: integer
        description: Classificação do comentário
      difficulty:
        type: integer
        description: Dificuldade do comentário
      comment:
        type: string
        description: Texto do comentário
      photos:
        type: array
        items:
          type: object
          properties:
            data:
              type: string
              format: byte
              description: Dados da foto codificados em base64
            contentType:
              type: string
              description: Tipo de conteúdo da foto (por exemplo, image/jpeg)
      userId:
        type: string
        description: Identificador do usuário que fez o comentário
      divingSpotId:
        type: string
        description: Identificador do ponto de mergulho relacionado ao comentário